+++ modules/authentication/portal-authentication-oauth/src/test/java/de/cuioss/portal/authentication/oauth/impl/Oauth2DiscoveryConfigurationProducerTest.java
@@ -17,8 +17,8 @@ package de.cuioss.portal.authentication.oauth.impl;
 
 import de.cuioss.portal.authentication.oauth.OAuthConfigKeys;
 import de.cuioss.portal.core.test.junit5.EnablePortalConfiguration;
-import de.cuioss.portal.core.test.junit5.mockwebserver.EnableMockWebServer;
-import de.cuioss.portal.core.test.junit5.mockwebserver.MockWebServerHolder;
+import de.cuioss.test.mockwebserver.EnableMockWebServer;
+import de.cuioss.test.mockwebserver.MockWebServerHolder;
 import de.cuioss.portal.core.test.mocks.configuration.PortalTestConfiguration;
 import de.cuioss.test.generator.impl.URLGenerator;
 import de.cuioss.test.juli.junit5.EnableTestLogger;
