+++ modules/test/portal-core-unit-testing/src/main/java/de/cuioss/test/mockwebserver/MockWebServerHolder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.cuioss.portal.core.test.junit5.mockwebserver;
+package de.cuioss.test.mockwebserver;
 
 import mockwebserver3.Dispatcher;
 import mockwebserver3.MockWebServer;
