+++ modules/test/portal-core-unit-testing/src/main/java/de/cuioss/test/mockwebserver/dispatcher/CombinedDispatcher.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.cuioss.portal.core.test.junit5.mockwebserver.dispatcher;
+package de.cuioss.test.mockwebserver.dispatcher;
 
+import de.cuioss.test.mockwebserver.EnableMockWebServer;
 import de.cuioss.tools.collect.CollectionLiterals;
 import de.cuioss.tools.logging.CuiLogger;
 import de.cuioss.tools.string.MoreStrings;
@@ -67,7 +68,7 @@ import java.util.List;
  *
  * @author Oliver Wolff
  * @see ModuleDispatcherElement
- * @see de.cuioss.portal.core.test.junit5.mockwebserver.EnableMockWebServer
+ * @see EnableMockWebServer
  * @since 1.0
  */
 @NoArgsConstructor
