+++ modules/test/portal-core-unit-testing/src/main/java/de/cuioss/test/mockwebserver/dispatcher/EndpointAnswerHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.cuioss.portal.core.test.junit5.mockwebserver.dispatcher;
+package de.cuioss.test.mockwebserver.dispatcher;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.Getter;
