+++ modules/test/portal-core-unit-testing/src/main/java/de/cuioss/test/mockwebserver/dispatcher/HttpMethodMapper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.cuioss.portal.core.test.junit5.mockwebserver.dispatcher;
+package de.cuioss.test.mockwebserver.dispatcher;
 
 import mockwebserver3.MockResponse;
 import mockwebserver3.RecordedRequest;
