+++ modules/test/portal-core-unit-testing/src/test/java/de/cuioss/test/mockwebserver/MockWebServerExtensionTest.java
@@ -13,16 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.cuioss.portal.core.test.junit5.mockwebserver;
+package de.cuioss.test.mockwebserver;
 
-import de.cuioss.portal.core.test.junit5.mockwebserver.dispatcher.BaseAllAcceptDispatcher;
-import de.cuioss.portal.core.test.junit5.mockwebserver.dispatcher.CombinedDispatcher;
+import de.cuioss.test.mockwebserver.dispatcher.BaseAllAcceptDispatcher;
+import de.cuioss.test.mockwebserver.dispatcher.CombinedDispatcher;
 import lombok.Getter;
 import lombok.Setter;
 import mockwebserver3.Dispatcher;
 import mockwebserver3.MockWebServer;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.io.IOException;
 import java.net.URI;
@@ -35,31 +34,24 @@ import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Test class for verifying that the {@link MockWebServerExtension} works correctly
- * when manually registered with {@link ExtendWith}.
+ * when using {@link EnableMockWebServer}.
  */
-@ExtendWith(MockWebServerExtension.class)
-@EnableMockWebServer(manualStart = true)
-class MockWebServerExtensionManualTest implements MockWebServerHolder {
+@EnableMockWebServer
+class MockWebServerExtensionTest implements MockWebServerHolder {
 
     @Getter
     @Setter
     private MockWebServer mockWebServer;
 
     @Test
-    void shouldProvideServerNotStartedServer() {
-        assertNotNull(mockWebServer, "Server should be injected even for manual start");
-        assertFalse(mockWebServer.getStarted(), "Server should not be started");
-        // Now start the server manually
-        assertDoesNotThrow(() -> mockWebServer.start());
-        assertTrue(mockWebServer.getStarted(), "Server should be started");
+    void shouldProvideServer() {
+        assertNotNull(mockWebServer);
+        assertTrue(mockWebServer.getStarted());
     }
 
     @Test
     void shouldHandleSimpleRequest() throws URISyntaxException, IOException, InterruptedException {
         String serverUrl = mockWebServer.url("/api").toString();
-
-        assertDoesNotThrow(() -> mockWebServer.start());
-
         HttpClient client = HttpClient.newHttpClient();
         HttpRequest request = HttpRequest.newBuilder()
                 .uri(new URI(serverUrl))
