+++ modules/test/portal-core-unit-testing/src/test/java/de/cuioss/test/mockwebserver/dispatcher/AllOkDispatcher.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.cuioss.portal.core.test.junit5.mockwebserver.dispatcher;
+package de.cuioss.test.mockwebserver.dispatcher;
 
 import lombok.NonNull;
 import mockwebserver3.MockResponse;
